+++ structures/hash_table_group_iterator.h
@@ -4,9 +4,11 @@
 
 #include "hash_table_group.h"
 
+#include <stdint.h>
+
 struct bsal_hash_table_group_iterator {
     struct bsal_hash_table_group *list;
-    int index;
+    uint64_t index;
     int size;
     int key_size;
     int value_size;
